requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
GetOldTweets3==0.0.11
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2==2.8.5
pycparser==2.20
pyquery==1.4.1
python-twitter==3.5
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
